requirements.txt
scrapy-1.3.2
pymongo-3.4.0
redis-2.10.5
cryptography-1.7.2
elasticsearch-5.2.0
kafka-1.3.2
Pillow-4.0.0
Django-1.10.5
BeautifulSoup4-4.5.3
pyasn1_modules-0.0.8